requirements.txt
azure-cognitiveservices-speech==1.37.0
openai==1.25.1
python-dotenv==1.0.1